Include fees handler in aHandler.
